Moved get_canary_id function into new utility file kayenta.py
